feat: enhance the view for replying chatroom messages (#79)
* chore: update dependency

* refactor: move the reply view to inside the VStack

* feat: add line limit when previewing replies

* feat: add ReplyPreviewView with stub sender and message

* feat: make text and reply preview appear leading

* feat: add vertical divider

* feat: add message view model to ReplyPreviewView

* feat: properly render replied message

* chore: improve MessageView_Previews

* feat: add extension for one-sided border

* feat: use one-sided border for vertical divider instead of rectangle

* refactor: move ReplyPreviewView to its own file

* feat: tap reply preview to jump up to reply

* feat: update ReplyPreviewView to take in custom border colour

The colour is independent of the message being displayed. This component
can also be reused when trying to reply message, which is why we attach
the rectangle border on this View instead of the MessageView

* feat: change scrolling anchor to center

* feat: scroll to message upon tapping from input bar reply preview

I don't quite like this code actually. It feels like I am asserting that
I have an existing messageBeingRepliedTo whenever I press on tap. There
are many states to maintain and I don't like it

* feat: enhance MessageInputBar reply preview design

* chore: reposition comment to prevent ambiguity

* chore: update note about center anchor point

* refactor: bundle the two Chatroom states in a single struct

* feat: use anchor bottom for onTap of message

* feat: add animation duration

I'm not sure if `easeIn` is even the correct Animation, but hey it seems
to work fine